Fix rendering of units when used as defaults of plugin keys
Which happened to `reportportal` plugin in the last release, and the
warning was not noticed in time. Doing it nicely called for adding a
custom Jinja test, which deserves its documentation, which deserves its
own section in our Jinja docs, which means renaming a couple of files
and other changes. But it looks nice now.
