+++ .gitignore
@@ -2,7 +2,7 @@
 
 /tmp/
 docs/code/autodocs/*.rst
-docs/code/template-filters.rst
+docs/code/template-extensions.rst
 docs/plugins/discover.rst
 docs/plugins/execute.rst
 docs/plugins/finish.rst
