+++ docs/code/index.rst
@@ -14,7 +14,7 @@ the documentation generated from sources linked below.
     :maxdepth: 2
 
     Class Overview <classes>
-    Template Filters <template-filters>
+    Template Extensions <template-extensions>
     Plugin Introduction <plugin-introduction>
     tmt <autodocs/tmt>
