+++ docs/templates/plugins.rst.j2
@@ -24,6 +24,8 @@
     Default: ``{{ actual_default }}``
         {% elif actual_default is integer %}
     Default: ``{{ actual_default }}``
+        {% elif actual_default is unit %}
+    Default: ``{{ actual_default }}``
         {% elif actual_default is none %}
     Default: *not set*
         {% elif actual_default is sequence %}
